Déplace TypesLieuEmploiFormation et lieu_emploi_ou_formation dans salarie.py
