+++ openfisca_france/model/prestations/agepi.py
@@ -2,7 +2,7 @@ from numpy import fabs, timedelta64
 
 from openfisca_france.model.base import Famille, Individu, Variable, Enum, MONTH, ADD,  \
     set_input_dispatch_by_period, set_input_divide_by_period, date, min_, not_
-from openfisca_france.model.revenus.activite.salarie import TypesContrat
+from openfisca_france.model.revenus.activite.salarie import TypesContrat, TypesLieuEmploiFormation
 
 
 class agepi_nbenf(Variable):
@@ -39,29 +39,6 @@ class duree_formation(Variable):
     label = "Durée de la formation en heures"
     definition_period = MONTH
     set_input = set_input_divide_by_period
-
-
-class TypesLieuEmploiFormation(Enum):
-    non_renseigne = "Non renseigné"
-    metropole = "Métropole"
-    guadeloupe = "Guadeloupe"
-    martinique = "Martinique"
-    guyane = "Guyane"
-    la_reunion = "La réunion"
-    saint_pierre_et_miquelon = "Saint Pierre et Miquelon"
-    mayotte = "Mayotte"
-    saint_bartelemy = "Saint Bartelemy"
-    saint_martin = "Saint Martin"
-
-
-class lieu_emploi_ou_formation(Variable):
-    value_type = Enum
-    possible_values = TypesLieuEmploiFormation
-    default_value = TypesLieuEmploiFormation.non_renseigne
-    entity = Individu
-    label = "Zone de l'emploi ou de la formation"
-    definition_period = MONTH
-    set_input = set_input_dispatch_by_period
 
 
 class TypesCategoriesDemandeurEmploi(Enum):
